requirements.txt
streamlit==1.44.1
openai==1.70.0
python-docx==1.1.2
weasyprint==65.0